requirements.txt
numpy
opencv-python-headless==4.5.3
pillow==8.3.1
pydub==0.25.0
moviepy==1.0.3
scikit-image